refactor: moves nav and footer to _app.js
